Rewrite building tests
