+++ api/__tests__/utilities/constants.ts
@@ -1,0 +1,17 @@
+export const badRequestResponse = {
+    message: "Bad Request",
+};
+
+export const forbiddenResponse = {
+    message: "Forbidden",
+};
+
+export const internalErrorResponse = {
+    detail: "An unknown error occurred",
+    message: "Internal Server Error",
+};
+
+export const notFoundResponse = {
+    detail: "Resource does not exist",
+    message: "Not Found",
+};
